Thinkstruct/requirements.txt
# Backend API
fastapi>=0.104.0
uvicorn>=0.24.0

# NLP & Machine Learning
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Database
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Cache
redis>=5.0.0

# Performance Testing
locust>=2.20.0
